move existing field boundary operations and fills and correct includes
