+++ amr-wind/boundary_conditions/field_boundary_fill/ABLFillMPL.cpp
@@ -1,6 +1,6 @@
 #include <utility>
 
-#include "amr-wind/wind_energy/ABLFillMPL.H"
+#include "amr-wind/boundary_conditions/field_boundary_fill/ABLFillMPL.H"
 
 namespace amr_wind {
