+++ amr-wind/boundary_conditions/field_boundary_fill/CMakeLists.txt
@@ -2,4 +2,8 @@ target_sources(${amr_wind_lib_name}
   PRIVATE
   OceanWavesFillInflow.cpp
   OceanWavesBoundary.cpp
+  ABLFillInflow.cpp
+  ABLBoundaryPlane.cpp
+  ABLFillMPL.cpp
+  ABLModulatedPowerLaw.cpp
   )
